[code publishing] Fix serialization problems, re-enable package download & secure w/ test
Package download (that is reconstructing a package from metadata on chain) failed because of a serialization issue. The issue is in getting `vector<u8>` as a field of a resource from Rust via the REST APIs back to Rust. Supposedly, one needs to define two structs `Foo` and `FooJson` in Rust of the same structure with different attributes (`serialize_as_string` for `u64`, `serde_bytes`). But I did not get this to work, so for now, I represent all binary data in a package as base64 compressed strings.

This PR also extends the `test_move_publish_flow` by a download at the end which verifies that the published package roundtrips well.

Morover, the PR adds a `move clean` command, in order to clean derived artifacts, including packages cached in `~/.move`.